refactor: ConversionForm component
