+++ .eslintignore
@@ -9,11 +9,4 @@ webpack.dev.ts
 jest.config.ts
 cypress/*
 cypress.config.ts
-
-**/test/*.ts
-**/test/*.tsx
-
-*.test.ts
-*.test.tsx
-
 __mocks__
